requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyJWT[crypto]==2.10.1
requests==2.32.3
python-dotenv==1.0.1
psutil==6.1.1
psycopg2-binary==2.9.10
